Favorites handling with react-query

+++ src/services/client/favorites.service.js
@@ -1,11 +1,13 @@
 import { API_URL } from '@/config/config'
 import { axiosClient } from '@/lib/utils/axios'
+import { isDev } from '@/lib/utils/isDev'
 
 class FavoritesService {
 	_FAVORITES = `${API_URL}/favorites`
 
 	/* ------------------------------ FavoritesIds ------------------------------ */
 	async getFavoritesIds(userId) {
+		if (!userId) return
 		const { data } = await axiosClient.get(this._FAVORITES, {
 			params: { userId }
 		})
@@ -14,6 +16,10 @@ class FavoritesService {
 	}
 	/* ------------------------------- AddFavorite ------------------------------ */
 	async addFavorite(userId, productId, category) {
+		if (!userId || !productId || !category) {
+			if (isDev()) console.error('Missing params')
+			return
+		}
 		await axiosClient.post(this._FAVORITES, {
 			userId,
 			productId,
@@ -22,6 +28,10 @@ class FavoritesService {
 	}
 	/* ----------------------------- DeleteFavorite ----------------------------- */
 	async deleteFavorite(userId, productId) {
+		if (!userId || !productId) {
+			if (isDev()) console.error('Missing params')
+			return
+		}
 		await axiosClient.delete(this._FAVORITES, {
 			params: { userId, productId }
 		})
@@ -29,6 +39,10 @@ class FavoritesService {
 
 	/* ---------------------------- DetailedFavorites --------------------------- */
 	async getDetailedFavorites(userId) {
+		if (!userId) {
+			if (isDev()) console.error('Missing userId')
+			return { data: [] }
+		}
 		const { data } = await axiosClient.get(`${this._FAVORITES}/detailed`, {
 			params: { userId }
 		})
